Changed DrawRing and DrawCircleSector angle params from int to float to allow greater accuracy. (#1656)

+++ examples/shapes/shapes_draw_ring.c
@@ -30,8 +30,8 @@ int main(void)
     float innerRadius = 80.0f;
     float outerRadius = 190.0f;
 
-    int startAngle = 0;
-    int endAngle = 360;
+    float startAngle = 0.0f;
+    float endAngle = 360.0f;
     int segments = 0;
 
     bool drawRing = true;
